+++ Services/Library.DAL/appsettings.json
@@ -1,0 +1,5 @@
+{
+  "ConnectionStrings": {
+    "default": "Data Source=(localdb)\\MSSQLLocalDB;Initial Catalog=BooksDB;Integrated Security=True"
+  }
+}
