+++ Services/Library.WcfService.Host/Program.cs
@@ -1,8 +1,7 @@
 ﻿using Library.DAL;
-using Library.WcfService.Host.HostService;
-using Microsoft.EntityFrameworkCore;
+
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
+
 using System.IO;
 using System.Threading.Tasks;
 
@@ -12,21 +11,14 @@ namespace Library.WcfService.Host
     {
         static async Task Main(string[] args)
         {
-            var dbInit = new DBInitializer();
-            await dbInit.Initialize(GetConnectionString());
+            var dbInitializer = new DBInitializer();
+            using (var db = dbInitializer.CreateDbContext(args))
+            {
+                var created = await db.Database.EnsureCreatedAsync();
+            }
 
             //var th = new TestHost();
             //await Task.Run(() => th.Initialize());
-        }
-
-        private static string GetConnectionString()
-        {
-            var configBuilder = new ConfigurationBuilder()
-                .SetBasePath(Directory.GetCurrentDirectory())
-                .AddJsonFile("appsettings.json");
-
-            var config = configBuilder.Build();
-            return config.GetConnectionString("default");
         }
     }
 }
